fix issues related to one-side-only methods
In rare cases, classes have methods that are present on both the client
and server, but a super class or interface declares that same method on
only one side. In such a case, only one side would get the mapping, and
when running the game on the other side, that method would still be
obfuscated.

This is fixed by also adding a mapping to a class, if any of its super
classes or interfaces implement the method on only one side.